QOLOE-380 move img to be relative to js/css and have it copied into dist/assets folder for reference
