+++ src/components/bs5/footer/footer.hbs
@@ -46,6 +46,7 @@
             <div class="col-xs-12 col-md-12 col-lg-3" aria-label="footer-navigation">
                 {{#if footerlinks}}
                 <div class="footer-column {{#if footerlinks.hasBorderColumn}}border-column{{/if}}">
+                    {{#if footerlinks.title}}<h3 class="footer-heading">{{footerlinks.title}}</h3>{{/if}}
                     <nav class="nav footer-link-list flex-column">
                         {{#each footerlinks.list }}
                         <a class="nav-link" href="{{ this.link }}" {{#if this.new_window }} target="_blank" {{/if}}>{{{ @key }}}</a>
