+++ src/components/bs5/fullPageWrapper/fullPage.stories.js
@@ -4,7 +4,7 @@ import Handlebars from "handlebars";
 
 import { dcTerms } from '../metaDcTerms/MetaDcTerms.data.json';
 import metaOpenGraphData from '../metaOpenGraph/MetaOpenGraph.data.json';
-import { masterbrand_variant } from '../header/header.data.json';
+import masterbrand_variant from '../header/header.variant.masterBrand.data.json';
 import { menu_state } from "../navbar/navbar.data.json";
 import breadcrumbsData from "../breadcrumbs/breadcrumbs.data.json";
 import searchData from "../searchInput/searchInput.data.json";
