+++ src/components/bs5/header/header.variant.masterBrand.data.json
@@ -1,0 +1,173 @@
+{
+  "icon-root": "assets/img/svg-icons.svg",
+  "icons": {
+    "menu-icon": "qld__icon__mobile-menu",
+    "search-icon": "qld__icon__search",
+    "close-icon": "qld__icon__close",
+    "chevron_down": "qld__icon__chevron-down"
+  },
+  "sitePreHeader": {
+    "url": {
+      "value": "https://qld.gov.au"
+    },
+    "text": {
+      "value": "www.qld.gov.au"
+    },
+    "CTA": [
+      {
+        "id": "CTAone",
+        "url": {
+          "value": "#"
+        },
+        "text": {
+          "value": "For Queenslanders"
+        },
+        "dropdown_enabled": true,
+        "dropdown_options": {
+          "dropdown_type": "list",
+          "view_more": true,
+          "label": "View all",
+          "url": "https://www.qld.gov.au/queenslanders",
+          "target": "_blank",
+          "dropdown_config": {
+            "groups": [
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Transport and motoring"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Employment and jobs"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Education and training"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Queensland and its government"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Health and wellbeing"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Community support"
+              }
+            ]
+          }
+        }
+      },
+      {
+        "id": "CTAtwo",
+        "url": {
+          "value": "#"
+        },
+        "text": {
+          "value": "Business and Industry"
+        },
+        "dropdown_enabled": true,
+        "dropdown_options": {
+          "dropdown_type": "list",
+          "view_more": true,
+          "label": "View all",
+          "url": "https://www.qld.gov.au/queenslanders",
+          "target": "_blank",
+          "dropdown_config": {
+            "groups": [
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Transport and motoring"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Employment and jobs"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Education and training"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Queensland and its government"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Health and wellbeing"
+              },
+              {
+                "action": "_blank",
+                "url": "#",
+                "label": "Community support"
+              }
+            ]
+          }
+        }
+      },
+      {
+        "id": "CTAthree",
+        "url": {
+          "value": "#"
+        },
+        "text": {
+          "value": "Login"
+        },
+        "dropdown_enabled": true,
+        "dropdown_options": {
+          "dropdown_type": "form",
+          "dropdown_config": {
+            "content": "<div></div>"
+          }
+        }
+      }
+    ]
+  },
+  "siteHeader": {
+    "config": {
+      "masterbrand_enabled": true,
+      "subbrand_enabled": false,
+      "cobrand_enabled": false,
+      "endorsed_enabled": false,
+      "standalone_enabled": false
+    },
+    "url": {
+      "value": "https://qld.gov.au"
+    },
+    "secondaryType": {
+      "logo": {
+        "value": ""
+      },
+      "siteTitle": {
+        "value": "Queensland Government Design System"
+      },
+      "subline": {
+        "value": "QGDS"
+      }
+    }
+  },
+  "siteSearchAsset": {
+    "value": "#",
+    "placeholder": "",
+    "label": "",
+    "formAction": {
+      "url": "https://www.forgov.qld.gov.au/search"
+    },
+    "options": {
+      "borderStyle": {
+        "full": true
+      }
+    }
+  }
+}
