+++ src/components/bs5/mainContainerWrapper/mainContainerWrapper.stories.js
@@ -3,7 +3,7 @@ import { MainContainerWrapperTest } from "./mainContainerWrapper.test.js";
 import init from "./../../../js/handlebars.init";
 import Handlebars from "handlebars";
 //Data
-import { masterbrand_variant } from '../header/header.data.json';
+import masterbrand_variant from '../header/header.variant.masterBrand.data.json';
 import { menu_state } from "../navbar/navbar.data.json";
 import breadcrumbsData from "../breadcrumbs/breadcrumbs.data.json";
 import searchData from "../searchInput/searchInput.data.json";
