+++ src/js/handlebars.partials.js
@@ -49,46 +49,46 @@ import video from "../components/bs5/video/video.hbs?raw";
  * @returns {void} Result of the helper operation
  */
 export default function handlebarsPartials(handlebars) {
-    handlebars.registerPartial("accordion", accordion);
-    handlebars.registerPartial("banner", banner);
-    handlebars.registerPartial("blockquote", blockquote);
-    handlebars.registerPartial("breadcrumbs", breadcrumbs);
-    handlebars.registerPartial("breadcrumbsWrapper", breadcrumbsWrapper);
-    handlebars.registerPartial("button", button);
-    handlebars.registerPartial("callout", callout);
-    handlebars.registerPartial("card", card);
-    handlebars.registerPartial("contentFooter", contentFooter);
-    handlebars.registerPartial("contentFooterWrapper", contentFooterWrapper);
-    handlebars.registerPartial("contentWrapper", contentWrapper);
-    handlebars.registerPartial("correctincorrect", correctincorrect);
-    handlebars.registerPartial("dateinput", dateinput);
-    handlebars.registerPartial("footer", footer);
-    handlebars.registerPartial("footerForgov", footerForgov);
-    handlebars.registerPartial("formcheck", formcheck);
-    handlebars.registerPartial("fullPage", fullPage);
-    handlebars.registerPartial("globalAlert", globalAlert);
-    handlebars.registerPartial("head", head);
-    handlebars.registerPartial("header", header);
-    handlebars.registerPartial("inpageAlert", inpageAlert);
-    handlebars.registerPartial("inpagenav", inpagenav);
-    handlebars.registerPartial("mainContainerWrapper", mainContainerWrapper);
-    handlebars.registerPartial("metaDcTerms", metaDcTerms);
-    handlebars.registerPartial("metaOpenGraph", metaOpenGraph);
-    handlebars.registerPartial("modal", modal);
-    handlebars.registerPartial("navbar", navbar);
-    handlebars.registerPartial("pagination", pagination);
-    handlebars.registerPartial("quickexit", quickexit);
-    handlebars.registerPartial("searchInput", searchInput);
-    handlebars.registerPartial("select", select);
-    handlebars.registerPartial("sidenav", sidenav);
-    handlebars.registerPartial("sidenavWrapper", sidenavWrapper);
-    handlebars.registerPartial("spinner", spinner);
-    handlebars.registerPartial("table", table);
-    handlebars.registerPartial("tag", tag);
-    handlebars.registerPartial("textarea", textarea);
-    handlebars.registerPartial("textbox", textbox);
-    handlebars.registerPartial("typography", typography);
-    handlebars.registerPartial("video", video);
+  handlebars.registerPartial("accordion", accordion);
+  handlebars.registerPartial("banner", banner);
+  handlebars.registerPartial("blockquote", blockquote);
+  handlebars.registerPartial("breadcrumbs", breadcrumbs);
+  handlebars.registerPartial("breadcrumbsWrapper", breadcrumbsWrapper);
+  handlebars.registerPartial("button", button);
+  handlebars.registerPartial("callout", callout);
+  handlebars.registerPartial("card", card);
+  handlebars.registerPartial("contentFooter", contentFooter);
+  handlebars.registerPartial("contentFooterWrapper", contentFooterWrapper);
+  handlebars.registerPartial("contentWrapper", contentWrapper);
+  handlebars.registerPartial("correctincorrect", correctincorrect);
+  handlebars.registerPartial("dateinput", dateinput);
+  handlebars.registerPartial("footer", footer);
+  handlebars.registerPartial("footerForgov", footerForgov);
+  handlebars.registerPartial("formcheck", formcheck);
+  handlebars.registerPartial("fullPage", fullPage);
+  handlebars.registerPartial("globalAlert", globalAlert);
+  handlebars.registerPartial("head", head);
+  handlebars.registerPartial("header", header);
+  handlebars.registerPartial("inpageAlert", inpageAlert);
+  handlebars.registerPartial("inpagenav", inpagenav);
+  handlebars.registerPartial("mainContainerWrapper", mainContainerWrapper);
+  handlebars.registerPartial("metaDcTerms", metaDcTerms);
+  handlebars.registerPartial("metaOpenGraph", metaOpenGraph);
+  handlebars.registerPartial("modal", modal);
+  handlebars.registerPartial("navbar", navbar);
+  handlebars.registerPartial("pagination", pagination);
+  handlebars.registerPartial("quickexit", quickexit);
+  handlebars.registerPartial("searchInput", searchInput);
+  handlebars.registerPartial("select", select);
+  handlebars.registerPartial("sidenav", sidenav);
+  handlebars.registerPartial("sidenavWrapper", sidenavWrapper);
+  handlebars.registerPartial("spinner", spinner);
+  handlebars.registerPartial("table", table);
+  handlebars.registerPartial("tag", tag);
+  handlebars.registerPartial("textarea", textarea);
+  handlebars.registerPartial("textbox", textbox);
+  handlebars.registerPartial("typography", typography);
+  handlebars.registerPartial("video", video);
 
 }
 if(typeof(Handlebars) !== 'undefined') {
